Split RdtEventManager into partial classes
