+++ SSMS EnvTabs/RdtEventManager.cs
@@ -7,17 +7,21 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Linq.Expressions;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace SSMS_EnvTabs
 {
-    internal sealed class RdtEventManager : IVsRunningDocTableEvents3, IVsSelectionEvents, IDisposable
+    internal sealed partial class RdtEventManager : IVsRunningDocTableEvents3, IVsSelectionEvents, IDisposable
     {
         private const uint SeidWindowFrame = 1;
         private const uint SeidDocumentFrame = 2;
         private const int RenameRetryCount = 20; // Increased to handle slow connection changes (up to 5s)
         private const int RenameRetryDelayMs = 250;
+        private static readonly Regex RenameEligibleRegex = new Regex(
+            @"^SQLQuery\d+\.sql\b",
+            RegexOptions.IgnoreCase | RegexOptions.CultureInvariant | RegexOptions.Compiled);
 
         private readonly AsyncPackage package;
         private readonly IVsRunningDocumentTable rdt;
@@ -149,11 +153,6 @@ namespace SSMS_EnvTabs
             }
         }
 
-        private void OnConfigRenamed(object sender, RenamedEventArgs e)
-        {
-            OnConfigChanged(sender, e);
-        }
-
         public void Dispose()
         {
             ThreadHelper.ThrowIfNotOnUIThread();
@@ -203,137 +202,6 @@ namespace SSMS_EnvTabs
             catch
             {
                 // best-effort
-            }
-        }
-
-        private void OnConfigChanged(object sender, FileSystemEventArgs e)
-        {
-            try
-            {
-                EnvTabsLog.Info($"Config file file system event: {e.ChangeType} - {e.FullPath}");
-                
-                // Debounce logic
-                CancellationToken token;
-                lock (debounceLock)
-                {
-                    debounceCts?.Cancel();
-                    debounceCts = new CancellationTokenSource();
-                    token = debounceCts.Token;
-                }
-
-                _ = package.JoinableTaskFactory.RunAsync(async () =>
-                {
-                    try
-                    {
-                        await Task.Delay(500, token);
-                        await package.JoinableTaskFactory.SwitchToMainThreadAsync(token);
-                        ReloadAndApplyConfig();
-                    }
-                    catch (OperationCanceledException)
-                    {
-                        // Ignored
-                    }
-                    catch (Exception ex)
-                    {
-                        EnvTabsLog.Info($"Config reload task failed: {ex.Message}");
-                    }
-                });
-            }
-            catch (Exception ex)
-            {
-                EnvTabsLog.Info($"OnConfigChanged error: {ex.Message}");
-            }
-        }
-
-        private void ReloadAndApplyConfig()
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-            EnvTabsLog.Info("Configuration changed, reloading...");
-
-            // Invalidate cache
-            cachedConfig = null;
-            // Clear suppressed "do not ask again" connections so that if user removed a rule, we effectively reset the "ignore" state
-            // and will prompt again if they connect to it.
-            AutoConfigurationService.ClearSuppressed();
-
-            var config = LoadConfigOrNull();
-            var rules = cachedRules; // LoadConfigOrNull updates cachedRules
-
-            if (config == null) return;
-
-            var docs = GetOpenDocumentsSnapshot();
-
-            // Rename
-            int renamedCount = 0;
-            foreach (var doc in docs)
-            {
-                try
-                {
-                     if (!string.IsNullOrWhiteSpace(doc.Server))
-                     {
-                         renamedCount += TabRenamer.ApplyRenamesOrThrow(new[] { (doc.Cookie, doc.Frame, doc.Server, doc.Database, doc.Caption) }, rules);
-                     }
-                }
-                catch (Exception ex)
-                {
-                    EnvTabsLog.Info($"Reload rename failed for {doc.Moniker}: {ex.Message}");
-                }
-            }
-
-            // Color
-            if (config.Settings?.EnableAutoColor == true)
-            {
-                try
-                {
-                    colorWriter.UpdateFromSnapshot(docs, rules);
-                }
-                catch (Exception ex)
-                {
-                     EnvTabsLog.Info($"Reload color update failed: {ex.Message}");
-                }
-            }
-        }
-
-        private TabGroupConfig LoadConfigOrNull()
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            try
-            {
-                string path = TabGroupConfigLoader.GetUserConfigPath();
-                if (!File.Exists(path))
-                {
-                    cachedConfig = null;
-                    cachedRules = null;
-                    cachedConfigLastWriteUtc = default;
-                    return null;
-                }
-
-                DateTime lastWriteUtc = File.GetLastWriteTimeUtc(path);
-                if (cachedConfig != null && lastWriteUtc == cachedConfigLastWriteUtc)
-                {
-                    return cachedConfig;
-                }
-
-                var loaded = TabGroupConfigLoader.LoadOrNull();
-                
-                if (loaded != null && loaded.Settings != null)
-                {
-                    EnvTabsLog.Enabled = loaded.Settings.EnableLogging;
-                }
-
-                cachedConfig = loaded;
-                cachedConfigLastWriteUtc = lastWriteUtc;
-                cachedRules = TabRuleMatcher.CompileRules(loaded);
-                return loaded;
-            }
-            catch (Exception ex)
-            {
-                EnvTabsLog.Info($"RdtEventManager: config reload failed: {ex.Message}");
-                cachedConfig = null;
-                cachedRules = null;
-                cachedConfigLastWriteUtc = default;
-                return null;
             }
         }
 
@@ -430,104 +298,6 @@ namespace SSMS_EnvTabs
             }
 
             return !needsRetry;
-        }
-
-        private static bool TryGetConnectionInfo(IVsWindowFrame frame, out string server, out string database)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            server = null;
-            database = null;
-
-            if (frame == null)
-            {
-                return false;
-            }
-
-            try
-            {
-                TryPopulateFromCaptions(frame, ref server, ref database);
-                return !string.IsNullOrWhiteSpace(server) || !string.IsNullOrWhiteSpace(database);
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        private static void TryPopulateFromCaptions(IVsWindowFrame frame, ref string server, ref string database)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-            if (frame == null) return;
-
-            string caption = TryReadFrameCaption(frame);
-            if (TryParseServerDatabaseFromCaption(caption, out string s1, out string d1))
-            {
-                if (string.IsNullOrWhiteSpace(server)) server = s1;
-                if (string.IsNullOrWhiteSpace(database)) database = d1;
-                return;
-            }
-
-            try
-            {
-                if (frame.GetProperty((int)__VSFPROPID.VSFPROPID_EditorCaption, out object editorCaptionObj) == VSConstants.S_OK)
-                {
-                    string editorCaption = editorCaptionObj as string;
-                    if (TryParseServerDatabaseFromCaption(editorCaption, out string s2, out string d2))
-                    {
-                        if (string.IsNullOrWhiteSpace(server)) server = s2;
-                        if (string.IsNullOrWhiteSpace(database)) database = d2;
-                    }
-                }
-            }
-            catch
-            {
-                // ignore
-            }
-        }
-
-        private static bool TryParseServerDatabaseFromCaption(string caption, out string server, out string database)
-        {
-            server = null;
-            database = null;
-
-            if (string.IsNullOrWhiteSpace(caption)) return false;
-
-            try
-            {
-                int dash = caption.IndexOf(" - ", StringComparison.Ordinal);
-                if (dash < 0)
-                {
-                    return false;
-                }
-
-                string tail = caption.Substring(dash + 3);
-                int paren = tail.IndexOf(" (", StringComparison.Ordinal);
-                if (paren >= 0)
-                {
-                    tail = tail.Substring(0, paren);
-                }
-
-                tail = tail.Trim();
-                if (string.IsNullOrWhiteSpace(tail)) return false;
-
-                int dot = tail.IndexOf('.');
-                if (dot > 0 && dot < tail.Length - 1)
-                {
-                    server = tail.Substring(0, dot).Trim();
-                    database = tail.Substring(dot + 1).Trim();
-                }
-                else
-                {
-                    server = tail;
-                }
-
-                return !string.IsNullOrWhiteSpace(server) || !string.IsNullOrWhiteSpace(database);
-            }
-            catch
-            {
-                return false;
-            }
         }
 
         private void TryHookDocViewConnectionEvents(uint docCookie, IVsWindowFrame frame)
@@ -669,156 +439,6 @@ namespace SSMS_EnvTabs
             });
         }
 
-        private void UpdateColorOnly(string reason)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            var config = LoadConfigOrNull();
-            if (config?.Settings?.EnableAutoColor != true)
-            {
-                return;
-            }
-
-            var rules = cachedRules ?? new List<TabRuleMatcher.CompiledRule>();
-            if (rules.Count == 0)
-            {
-                return;
-            }
-
-            try
-            {
-                var docs = GetOpenDocumentsSnapshot();
-                colorWriter.UpdateFromSnapshot(docs, rules);
-            }
-            catch (Exception ex)
-            {
-                EnvTabsLog.Info($"ColorByRegex update failed ({reason}): {ex.Message}");
-            }
-        }
-
-        private List<OpenDocumentInfo> GetOpenDocumentsSnapshot()
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            var list = new List<OpenDocumentInfo>();
-
-            rdt.GetRunningDocumentsEnum(out IEnumRunningDocuments enumDocs);
-            if (enumDocs == null)
-            {
-                return list;
-            }
-
-            uint[] cookies = new uint[1];
-            while (enumDocs.Next(1, cookies, out uint fetched) == VSConstants.S_OK && fetched == 1)
-            {
-                uint cookie = cookies[0];
-                string moniker = TryGetMonikerFromCookie(cookie);
-                if (string.IsNullOrWhiteSpace(moniker))
-                {
-                    continue;
-                }
-
-                IVsWindowFrame frame = TryGetFrameFromMoniker(moniker);
-                if (frame == null)
-                {
-                    continue;
-                }
-
-                string caption = TryReadFrameCaption(frame);
-                TryGetConnectionInfo(frame, out string server, out string database);
-                list.Add(new OpenDocumentInfo
-                {
-                    Cookie = cookie,
-                    Frame = frame,
-                    Caption = caption,
-                    Moniker = moniker,
-                    Server = server,
-                    Database = database
-                });
-            }
-
-            return list;
-        }
-
-        private string TryGetMonikerFromCookie(uint docCookie)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (docCookie == 0) return null;
-
-            try
-            {
-                rdt.GetDocumentInfo(
-                    docCookie,
-                    out uint _,
-                    out uint _,
-                    out uint _,
-                    out string moniker,
-                    out IVsHierarchy _,
-                    out uint _,
-                    out IntPtr _);
-
-                return moniker;
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        private static string TryReadFrameCaption(IVsWindowFrame frame)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (frame == null) return null;
-
-            try
-            {
-                if (frame.GetProperty((int)__VSFPROPID.VSFPROPID_Caption, out object caption) == VSConstants.S_OK)
-                {
-                    return caption as string;
-                }
-            }
-            catch
-            {
-                // ignore
-            }
-
-            return null;
-        }
-
-        private static string TryReadFrameEditorCaption(IVsWindowFrame frame)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (frame == null) return null;
-
-            try
-            {
-                if (frame.GetProperty((int)__VSFPROPID.VSFPROPID_EditorCaption, out object caption) == VSConstants.S_OK)
-                {
-                    return caption as string;
-                }
-            }
-            catch
-            {
-                // ignore
-            }
-
-            return null;
-        }
-
-        private void LogFrameCaptions(IVsWindowFrame frame, string reason)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (frame == null) return;
-
-            string caption = TryReadFrameCaption(frame) ?? string.Empty;
-            string editorCaption = TryReadFrameEditorCaption(frame) ?? string.Empty;
-            EnvTabsLog.Info($"Frame caption ({reason}): '{caption}' | editor='{editorCaption}'");
-        }
-
         private void ScheduleRenameRetry(uint docCookie, string reason)
         {
             if (docCookie == 0) return;
@@ -873,237 +493,5 @@ namespace SSMS_EnvTabs
             });
         }
 
-        private bool TryGetMonikerFromFrame(IVsWindowFrame frame, out string moniker)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            moniker = null;
-            if (frame == null) return false;
-
-            try
-            {
-                if (frame.GetProperty((int)__VSFPROPID.VSFPROPID_pszMkDocument, out object mk) == VSConstants.S_OK)
-                {
-                    moniker = mk as string;
-                    return !string.IsNullOrWhiteSpace(moniker);
-                }
-            }
-            catch
-            {
-                // ignore
-            }
-
-            return false;
-        }
-
-        private bool TryGetCookieFromMoniker(string moniker, out uint cookie)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            cookie = 0;
-            if (string.IsNullOrWhiteSpace(moniker)) return false;
-
-            try
-            {
-                rdt.FindAndLockDocument((uint)_VSRDTFLAGS.RDT_NoLock, moniker, out IVsHierarchy _, out uint _, out IntPtr _, out cookie);
-                return cookie != 0;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        private IVsWindowFrame TryGetFrameFromMoniker(string moniker)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (shellOpenDoc == null || string.IsNullOrWhiteSpace(moniker)) return null;
-
-            try
-            {
-                Guid logicalView = Guid.Empty;
-                shellOpenDoc.OpenDocumentViaProject(
-                    moniker,
-                    ref logicalView,
-                    out Microsoft.VisualStudio.OLE.Interop.IServiceProvider _,
-                    out IVsUIHierarchy _,
-                    out uint _,
-                    out IVsWindowFrame frame);
-
-                return frame;
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        // --- RDT events ---
-        public int OnAfterFirstDocumentLock(uint docCookie, uint dwRDTLockType, uint dwReadLocksRemaining, uint dwEditLocksRemaining)
-        {
-            return VSConstants.S_OK;
-        }
-
-        public int OnAfterDocumentWindowShow(uint docCookie, int fFirstShow, IVsWindowFrame pFrame)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (pFrame != null)
-            {
-                TryHookDocViewConnectionEvents(docCookie, pFrame);
-                bool done = HandlePotentialChange(docCookie, pFrame, reason: "DocumentWindowShow");
-                if (!done)
-                {
-                    ScheduleRenameRetry(docCookie, "DocumentWindowShow");
-                }
-            }
-
-            return VSConstants.S_OK;
-        }
-
-        public int OnAfterAttributeChangeEx(uint docCookie, uint grfAttribs, IVsHierarchy pHierOld, uint itemidOld, string pszMkDocumentOld, IVsHierarchy pHierNew, uint itemidNew, string pszMkDocumentNew)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            string moniker = pszMkDocumentNew;
-            if (string.IsNullOrWhiteSpace(moniker))
-            {
-                moniker = pszMkDocumentOld;
-            }
-
-            if (!string.IsNullOrWhiteSpace(moniker))
-            {
-                IVsWindowFrame frame = TryGetFrameFromMoniker(moniker);
-                if (frame != null)
-                {
-                    HandlePotentialChange(docCookie, frame, reason: "AttributeChangeEx");
-                }
-            }
-
-            return VSConstants.S_OK;
-        }
-
-        public int OnBeforeLastDocumentUnlock(uint docCookie, uint dwRDTLockType, uint dwReadLocksRemaining, uint dwEditLocksRemaining)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (dwReadLocksRemaining == 0 && dwEditLocksRemaining == 0)
-            {
-                TabRenamer.ForgetCookie(docCookie);
-                renameRetryCounts.Remove(docCookie);
-
-                if (docViewSubscriptionsByCookie.TryGetValue(docCookie, out var subs))
-                {
-                    foreach (var sub in subs)
-                    {
-                        try { sub.Dispose(); } catch { }
-                    }
-                    docViewSubscriptionsByCookie.Remove(docCookie);
-                }
-
-                UpdateColorOnly("LastDocumentUnlock");
-            }
-
-            return VSConstants.S_OK;
-        }
-
-        public int OnAfterDocumentWindowHide(uint docCookie, IVsWindowFrame pFrame) => VSConstants.S_OK;
-        public int OnAfterSave(uint docCookie) 
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-            UpdateColorOnly("AfterSave");
-            return VSConstants.S_OK;
-        }
-        public int OnBeforeSave(uint docCookie) => VSConstants.S_OK;
-        public int OnAfterAttributeChange(uint docCookie, uint grfAttribs)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            string moniker = TryGetMonikerFromCookie(docCookie);
-            if (!string.IsNullOrWhiteSpace(moniker))
-            {
-                IVsWindowFrame frame = TryGetFrameFromMoniker(moniker);
-                if (frame != null)
-                {
-                    bool done = HandlePotentialChange(docCookie, frame, reason: "AttributeChange");
-                    
-                    // Force a retry if it's an AttributeChange, because the caption is likely stale.
-                    if (!done)
-                    {
-                        ScheduleRenameRetry(docCookie, "AttributeChange");
-                    }
-                    else
-                    {
-                        ScheduleRenameRetry(docCookie, "AttributeChange:Force");
-                    }
-                }
-            }
-
-            return VSConstants.S_OK;
-        }
-        public int OnBeforeDocumentWindowShow(uint docCookie, int fFirstShow, IVsWindowFrame pFrame) => VSConstants.S_OK;
-
-        // --- Selection events ---
-
-        public int OnElementValueChanged(uint elementid, object varValueOld, object varValueNew)
-        {
-            ThreadHelper.ThrowIfNotOnUIThread();
-
-            if (elementid == SeidDocumentFrame || elementid == SeidWindowFrame)
-            {
-                if (varValueNew is IVsWindowFrame frame)
-                {
-                    uint cookie = 0;
-                    string moniker = null;
-                    if (TryGetMonikerFromFrame(frame, out moniker))
-                    {
-                        TryGetCookieFromMoniker(moniker, out cookie);
-                    }
-                    
-                    if (cookie == 0)
-                    {
-                         // Found frame but no cookie
-                    }
-                    else
-                    {
-                        TryHookDocViewConnectionEvents(cookie, frame);
-                    }
-
-                    bool done = HandlePotentialChange(cookie, frame, reason: "ActiveFrameChanged");
-                    if (!done && cookie != 0)
-                    {
-                        ScheduleRenameRetry(cookie, "ActiveFrameChanged");
-                    }
-                }
-            }
-
-            return VSConstants.S_OK;
-        }
-
-        public int OnCmdUIContextChanged(uint dwCmdUICookie, int fActive) => VSConstants.S_OK;
-
-        public int OnSelectionChanged(IVsHierarchy pHierOld, uint itemidOld, IVsMultiItemSelect pMISOld, ISelectionContainer pSCOld,
-            IVsHierarchy pHierNew, uint itemidNew, IVsMultiItemSelect pMISNew, ISelectionContainer pSCNew) => VSConstants.S_OK;
-
-        private static bool IsRenameEligible(string caption)
-        {
-            try
-            {
-                if (string.IsNullOrWhiteSpace(caption))
-                {
-                    return false;
-                }
-
-                return System.Text.RegularExpressions.Regex.IsMatch(
-                    caption,
-                    @"^SQLQuery\d+\.sql\b",
-                    System.Text.RegularExpressions.RegexOptions.IgnoreCase | System.Text.RegularExpressions.RegexOptions.CultureInvariant);
-            }
-            catch
-            {
-                return false;
-            }
-        }
     }
 }
