+++ SSMS EnvTabs/SSMS EnvTabs.csproj
@@ -54,6 +54,9 @@
     </Compile>
     <Compile Include="OpenConfigCommand.cs" />
     <Compile Include="RdtEventManager.cs" />
+    <Compile Include="RdtEventManager.Config.cs" />
+    <Compile Include="RdtEventManager.Documents.cs" />
+    <Compile Include="RdtEventManager.Events.cs" />
     <Compile Include="SSMS_EnvTabsPackage.cs" />
     <Compile Include="TabGroupConfigLoader.cs" />
     <Compile Include="TabGroupConfigModels.cs" />
